requirements.txt
easydict==1.10
matplotlib==3.7.1
numpy==1.23.1
Pillow==9.2.0
PyYAML==6.0
randaugment==1.0.2
scikit_learn==1.2.2
timm==0.9.2
torch==1.12.1
torchmetrics==0.11.4
torchvision==0.13.1
tqdm==4.64.1